@@ -33,16 +33,21 @@
 	},
 	"scripts": {
 		"all": "npm run travis && npm run pack-and-deploy",
-		"clean": "rm -rf lib && rm -rf dist && rm -rf node_modules && rm -f package-lock.json",
+		"clean-offline": "rm -rf coverage && rm -rf dist",
+		"clean": "npm run clean-offline && rm -rf node_modules && rm -f package-lock.json",
+		"build-and-test": "npm run prettiercheck && npm run lint && npm run build && npm test",
 		"prettiercheck": "prettier --check \"src/**/*.ts\" \"test/**/*.ts\"",
 		"format": "prettier --write \"src/**/*.ts\" \"test/**/*.ts\"",
-		"build": "tsc && mkdir dist && uglifyjs --compress --mangle -o dist/main.min.js -- lib/main.js && cp src/main.d.ts dist/",
-		"test": "jest --config jestconfig.json",
-		"travis": "npm run clean && npm i && npm run prettiercheck && npm run build && npm test",
+		"lint": "eslint . --ext .js --ext .ts",
+		"build": "rm -rf lib; tsc && mkdir dist && uglifyjs --compress --mangle -o dist/main.min.js -- lib/main.js && cp src/main.d.ts dist/",
+		"test": "jest",
+		"travis": "npm run clean && npm i && npm run build-and-test",
 		"pack-and-deploy": "npm pack && eval mv $npm_package_name-$npm_package_version.tgz $(echo $npm_package_config_deploy_target_dir)/",
 		"version": "git add -A src",
 		"postversion": "git push && git push --tags",
+		"re": "npm run format && npm run clean-offline && npm run build-and-test && npm run pack-and-deploy",
 		"loc": "git ls-files | grep -v package-lock.json | tr -d '\\r' | tr '\\n' '\\0' | xargs -0 wc -l",
+		"f": "npm run format && npm run all",
 		"start": "npm run all"
 	},
 	"config": {
@@ -51,17 +56,22 @@
 	"publishConfig": {},
 	"dependencies": {},
 	"devDependencies": {
-		"@d3fc/d3fc-technical-indicator": "^8.0.0",
-		"@types/jest": "^26.0.0",
-		"@types/node": "^14.0.13",
-		"d3-array": "^2.4.0",
-		"jest": "^26.0.1",
+		"@babel/core": "^7.11.6",
+		"@babel/preset-env": "^7.11.5",
+		"@babel/preset-typescript": "^7.10.4",
+		"@d3fc/d3fc-technical-indicator": "^8.0.1",
+		"@types/eslint": "^7.2.2",
+		"@types/jest": "^26.0.13",
+		"@types/node": "^14.10.1",
+		"@typescript-eslint/eslint-plugin": "^4.1.1",
+		"@typescript-eslint/parser": "^4.1.1",
+		"d3-array": "^2.7.1",
+		"eslint": "^7.9.0",
+		"jest": "^26.4.2",
 		"macd": "^1.0.2",
-		"prettier": "^2.0.5",
-		"ts-jest": "^26.1.0",
-		"tslint": "^6.1.2",
-		"tslint-config-prettier": "^1.18.0",
-		"typescript": "^3.9.5",
+		"prettier": "^2.1.1",
+		"semver": "^7.3.2",
+		"typescript": "^4.0.2",
 		"uglify-js": "^3.9.4"
 	},
 	"peerDependencies": {},
